core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/component/cluster/page/clusterpositionmapbucket/OClusterPositionMapBucketRemoveOperation.java
package com.orientechnologies.orient.core.storage.impl.local.paginated.wal.component.cluster.page.clusterpositionmapbucket;

import com.orientechnologies.common.serialization.types.OIntegerSerializer;
import com.orientechnologies.common.serialization.types.OLongSerializer;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OLogSequenceNumber;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OPageOperation;
import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.WALRecordTypes;

import java.nio.ByteBuffer;

public class OClusterPositionMapBucketRemoveOperation extends OPageOperation {
  private int  recordIndex;
  private int  recordPosition;
  private long recordPageIndex;

  public OClusterPositionMapBucketRemoveOperation() {
  }

  public OClusterPositionMapBucketRemoveOperation(OLogSequenceNumber pageLSN, long fileId, long pageIndex, int recordIndex,
      int recordPosition, long recordPageIndex) {
    super(pageLSN, fileId, pageIndex);

    this.recordIndex = recordIndex;
    this.recordPosition = recordPosition;
    this.recordPageIndex = recordPageIndex;
  }

  @Override
  public byte getId() {
    return WALRecordTypes.CLUSTER_POSITION_MAP_BUCKET_REMOVE_OPERATION;
  }

  @Override
  public int serializedSize() {
    return super.serializedSize() + 2 * OIntegerSerializer.INT_SIZE + OLongSerializer.LONG_SIZE;
  }

  @Override
  public int toStream(byte[] content, int offset) {
    offset = super.toStream(content, offset);

    OIntegerSerializer.INSTANCE.serializeNative(recordIndex, content, offset);
    offset += OIntegerSerializer.INT_SIZE;

    OIntegerSerializer.INSTANCE.serializeNative(recordPosition, content, offset);
    offset += OIntegerSerializer.INT_SIZE;

    OLongSerializer.INSTANCE.serializeNative(recordPageIndex, content, offset);
    offset += OLongSerializer.LONG_SIZE;

    return offset;
  }

  @Override
  public void toStream(ByteBuffer buffer) {
    super.toStream(buffer);

    buffer.putInt(recordIndex);
    buffer.putInt(recordPosition);
    buffer.putLong(recordPageIndex);
  }

  @Override
  public int fromStream(byte[] content, int offset) {
    offset = super.fromStream(content, offset);

    recordIndex = OIntegerSerializer.INSTANCE.deserializeNative(content, offset);
    offset += OIntegerSerializer.INT_SIZE;

    recordPosition = OIntegerSerializer.INSTANCE.deserializeNative(content, offset);
    offset += OIntegerSerializer.INT_SIZE;

    recordPageIndex = OLongSerializer.INSTANCE.deserializeNative(content, offset);
    offset += OLongSerializer.LONG_SIZE;

    return offset;
  }

  public int getRecordIndex() {
    return recordIndex;
  }

  public int getRecordPosition() {
    return recordPosition;
  }

  public long getRecordPageIndex() {
    return recordPageIndex;
  }
}
